copyright/pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "doccollate-copyright"
version = "0.1.0"
description = "Standalone copyright form generator for DocCollate"
requires-python = ">=3.10"
dependencies = [
  "docxtpl",
  "jieba",
  "openai",
  "openpyxl",
  "pdfplumber",
  "pydantic>=2,<3",
  "pydantic-ai",
  "python-docx",
  "python-dotenv",
  "pyyaml",
  "rank-bm25",
  "tomli; python_version < '3.11'",
]

[project.scripts]
doccollate-copyright = "doccollate_copyright.cli:main"

[tool.hatch.build]
include = [
  "src/doccollate_copyright/resources/*",
]

[tool.hatch.build.targets.wheel]
packages = ["src/doccollate_copyright"]

[tool.doccollate.llm]
model = "qwen3-max"
skeleton_model = "qwen3-max"
final_model = "qwen3-max"

[tool.doccollate.templates]
copyright = "src/doccollate_copyright/resources/software_copyright_application_form.docx"

[tool.doccollate.doccollate]
contact_info = "src/doccollate_copyright/resources/soft_copyright.yaml"
preset_choice = ""

[tool.doccollate.dates]
assess_completion_days_ago = 14
assess_dev_months_ago = 5
